added class for handling and parsing time data from the RTC to display on the lamps. Changed all uint_least8_t variables to unsigned int due to error with division

+++ src/sensors.cpp
@@ -5,24 +5,24 @@ void Sensors::setupSensors(){
     MQ9.init();
 }
 
-std::pair<uint_least8_t, uint_least8_t> Sensors::getTemperature(){
+std::pair<unsigned int, unsigned int> Sensors::getTemperature(){
     float temperature = dht.readTemperature();
-    uint_least8_t tempTens = (uint_least8_t)(temperature / 10);
+    unsigned int tempTens = (unsigned int)(temperature / 10);
     if(tempTens > 9){
         tempTens = 9;
     }
-    uint_least8_t tempOnes = (int)temperature % 10;
+    unsigned int tempOnes = (int)temperature % 10;
 
-    return std::pair<uint_least8_t, uint_least8_t>(tempTens, tempOnes);
+    return std::pair<unsigned int, unsigned int>(tempTens, tempOnes);
 } 
 
-std::pair<uint_least8_t, uint_least8_t> Sensors::getHumidity(){
+std::pair<unsigned int, unsigned int> Sensors::getHumidity(){
     float humidity = dht.readHumidity();
-    uint_least8_t humTens = (uint_least8_t)(humidity / 10);
+    unsigned int humTens = (unsigned int)(humidity / 10);
     if(humTens > 9){
         humTens = 9;
     }
-    uint_least8_t humOnes = (int)humidity % 10;
+    unsigned int humOnes = (int)humidity % 10;
     
-    return std::pair<uint_least8_t, uint_least8_t>(humTens, humOnes);
+    return std::pair<unsigned int, unsigned int>(humTens, humOnes);
 }
